+++ CHANGELOG.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.2.1] - 2022-06-17
+## [0.6.0] - 2022-06-17
 
 - Fix style in `link` component.
+
+## [0.5.0] - 2022-06-17
+
+- Completed `Loader` component.
+
+## [0.3.0] - 2022-06-16
+
+- Completed `Tabs` component. Added loading tab content on demand.
 
 ## [0.2.0] - 2022-06-15